/trunk, sawRobotIO1394: saRobot1394, fixed display for board id in exception message
git-svn-id: https://svn.lcsr.jhu.edu/cisst/trunk/saw/components/sawRobotIO1394@4451 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ code/osaRobot1394.cpp
@@ -215,7 +215,7 @@ void osaRobot1394::PollValidity(void)
              board != UniqueBoards_.end();
              ++board) {
             if (!board->second->ValidRead()) {
-                message << board->second->GetBoardId() << " ";
+                message << static_cast<int>(board->second->GetBoardId()) << " ";
             }
         }
         cmnThrow(osaRuntimeError1394(message.str()));
